Move provider to same file as reconciler
For consistency with Serving.
