+++ pkg/reconciler/v1alpha1/subscription/subscription.go
@@ -19,7 +19,13 @@ package subscription
 import (
 	"context"
 	"fmt"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/record"
 	"net/url"
+	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"sigs.k8s.io/controller-runtime/pkg/source"
 
 	"github.com/golang/glog"
 	eventingduck "github.com/knative/eventing/pkg/apis/duck/v1alpha1"
@@ -41,6 +47,10 @@ import (
 )
 
 const (
+	// controllerAgentName is the string used by this controller to identify
+	// itself when creating events.
+	controllerAgentName = "subscription-controller"
+
 	finalizerName = controllerAgentName
 
 	// Name of the corev1.Events emitted from the reconciliation process
@@ -51,6 +61,37 @@ const (
 	subscriberResolveFailed        = "SubscriberResolveFailed"
 	resultResolveFailed            = "ResultResolveFailed"
 )
+
+type reconciler struct {
+	client        client.Client
+	restConfig    *rest.Config
+	dynamicClient dynamic.Interface
+	recorder      record.EventRecorder
+}
+
+// Verify the struct implements reconcile.Reconciler
+var _ reconcile.Reconciler = &reconciler{}
+
+// ProvideController returns a Subscription controller.
+func ProvideController(mgr manager.Manager) (controller.Controller, error) {
+	// Setup a new controller to Reconcile Subscriptions.
+	c, err := controller.New(controllerAgentName, mgr, controller.Options{
+		Reconciler: &reconciler{
+			recorder: mgr.GetRecorder(controllerAgentName),
+		},
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	// Watch Subscription events and enqueue Subscription object key.
+	if err := c.Watch(&source.Kind{Type: &v1alpha1.Subscription{}}, &handler.EnqueueRequestForObject{}); err != nil {
+		return nil, err
+	}
+
+	return c, nil
+}
+
 
 // Reconcile compares the actual state with the desired, and attempts to
 // converge the two. It then updates the Status block of the Subscription resource
@@ -446,4 +487,17 @@ func removeFinalizer(sub *v1alpha1.Subscription) {
 	finalizers := sets.NewString(sub.Finalizers...)
 	finalizers.Delete(finalizerName)
 	sub.Finalizers = finalizers.List()
+}
+
+
+func (r *reconciler) InjectClient(c client.Client) error {
+	r.client = c
+	return nil
+}
+
+func (r *reconciler) InjectConfig(c *rest.Config) error {
+	r.restConfig = c
+	var err error
+	r.dynamicClient, err = dynamic.NewForConfig(c)
+	return err
 }
